requirements.txt
Django==2.2.4
django-cors-headers==3.5.0
django-filter==2.2.0
djangorestframework==3.10.2
gunicorn==20.0.4
Markdown==3.1.1
psycopg2-binary==2.8.4
pytz==2019.2
sqlparse==0.3.0
